@@ -1,7 +1,8 @@
 package com.example.DemoGraphQL.repository;
 
 import com.example.DemoGraphQL.model.Author;
-import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorRepository extends CrudRepository<Author, Long> {
+public interface AuthorRepository extends JpaRepository<Author, Long>
+{
 }
