@@ -1,7 +1,7 @@
 package com.example.DemoGraphQL.repository;
 
 import com.example.DemoGraphQL.model.Book;
-import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends CrudRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Long> {
 }
