@@ -1,9 +1,9 @@
 package com.example.DemoGraphQL.resolver;
 
-import com.coxautodev.graphql.tools.GraphQLResolver;
 import com.example.DemoGraphQL.model.Author;
 import com.example.DemoGraphQL.model.Book;
 import com.example.DemoGraphQL.repository.AuthorRepository;
+import graphql.kickstart.tools.GraphQLResolver;
 
 public class BookResolver implements GraphQLResolver<Book> {
     private AuthorRepository authorRepository;
@@ -13,6 +13,8 @@ public class BookResolver implements GraphQLResolver<Book> {
     }
 
     public Author getAuthor(Book book) {
-        return authorRepository.findOne(book.getAuthor().getId());
+        return authorRepository.findById(book.getAuthor()
+                                             .getId())
+                               .orElseThrow(null);
     }
 }
